requirements.txt
Flask==2.1.0
Flask-PyMongo==2.3.0
python-dotenv==0.19.0
Werkzeug==2.1.0